Let designation trigger load
Designating a default now also loads the new default as the current
resource.  This is to prevent some counter-intuitive use cases.  E.g.
before this commit, changing the default the profile and then changing the
default username in a single command changes the default username of the
old profile since the old profile is still loaded.